Interface fixação frete

+++ 0x08/resumo_pedido/Pedido.java
@@ -10,7 +10,7 @@ public class Pedido {
     private Frete frete;
 
 
-    public Pedido(int codigo, int peso, double total) {
+    public Pedido(int codigo, double peso, double total) {
         this.codigo = codigo;
         this.peso = peso;
         this.total = total;
@@ -28,7 +28,7 @@ public class Pedido {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(double peso) {
         this.peso = peso;
     }
